쉐이더 설정 및 수정, 203120-205220프리팹 수정, Json 추가 및 수정

+++ Assets/Scripts/Utility/Interaction/InteractionEvent.cs
@@ -75,7 +75,7 @@ namespace Utility.Interaction
         {
             public InteractionObject interactObj;
             public int index;
-            public int waitSeconds;
+            public float waitSeconds;
             public bool useFadeOut;
         }
